+++ src/main/java/br/com/zalf/prolog/webservice/v3/frota/pneu/pneuservico/PneuServicoCadastroDao.java
@@ -9,5 +9,5 @@ import org.springframework.stereotype.Repository;
  * @author Guilherme Steinert (https://github.com/steinert999)
  */
 @Repository
-public interface PneuServicoCadastroV3Dao extends JpaRepository<PneuServicoCadastroEntity, Long> {
+public interface PneuServicoCadastroDao extends JpaRepository<PneuServicoCadastroEntity, Long> {
 }