+++ src/main/java/br/com/zalf/prolog/webservice/v3/frota/pneu/pneuservico/PneuServicoRealizadoDao.java
@@ -9,5 +9,5 @@ import org.springframework.stereotype.Repository;
  * @author Guilherme Steinert (https://github.com/steinert999)
  */
 @Repository
-public interface PneuServicoRealizadoV3Dao extends JpaRepository<PneuServicoRealizadoEntity, Long> {
+public interface PneuServicoRealizadoDao extends JpaRepository<PneuServicoRealizadoEntity, Long> {
 }