+++ src/main/java/br/com/zalf/prolog/webservice/v3/frota/veiculo/VeiculoDao.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface VeiculoV3Dao extends JpaRepository<VeiculoEntity, Long> {
+public interface VeiculoDao extends JpaRepository<VeiculoEntity, Long> {
 }